+++ flaskr/__init__.py
@@ -1,1 +1,4 @@
-from flaskr import app,init_db
+
+from config import app
+from db import init_db
+from flaskr_routes import *
